Implemented map in lab1, axios in lab2
