+++ lebedkinAnatoliy/.eslintrc.js
@@ -1,4 +1,5 @@
 module.exports = {
   root: true,
   extends: '@react-native-community',
+  parser: '@babel/eslint-parser',
 };
