+++ lebedkinAnatoliy/package.json
@@ -12,15 +12,18 @@
   "dependencies": {
     "@react-navigation/bottom-tabs": "^6.4.0",
     "@react-navigation/native": "^6.0.13",
+    "axios": "^1.1.3",
     "react": "17.0.2",
     "react-native": "0.68.2",
     "react-native-gesture-handler": "^2.7.0",
     "react-native-safe-area-context": "^4.4.1",
     "react-native-screens": "^3.18.0",
+    "react-native-vector-icons": "^9.2.0",
     "react-native-webview": "^11.23.1"
   },
   "devDependencies": {
     "@babel/core": "7.19.0",
+    "@babel/eslint-parser": "^7.19.1",
     "@babel/runtime": "7.19.0",
     "@react-native-community/eslint-config": "2.0.0",
     "babel-jest": "26.6.3",
